Restructure tests [#163984766]
